+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.24.1
 
 * minor pylint improvements
+* better coverage
+* SyncableDroppable.cleanup() bugs out, wrote a quick patch for it to do nothing and filed as #61.